+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - #2624 Added "Maximum holding time" setting to services and analyses
+- #2639 Fix sampletype-related indexes for AnalysisSpec type are not indexed
+- #2638 Fix AttributeError on upgrade step 2654 (reindex_getDueDate)
 - #2569 Fix samples are indicated as late when Turnaround Time is zero
 - #2636 Fix JS Error in WS Template edit form
 - #2635 Remove reindexing of Analyses and Analysis Services on Category change